adding webpack
reconfigured folder structure
