+++ .babelrc
@@ -1,5 +1,16 @@
 {
   "presets": [
-    "env"
+    [
+      "@babel/preset-env",
+      {
+        "useBuiltIns": "usage",
+        "targets": {
+          "browsers": [
+            "last 2 versions",
+            "not ie < 11"
+          ]
+        }
+      }
+    ]
   ]
 }
