--- README.md
@@ -1,8 +1,0 @@
-1. create database
-2. import base db from /sql
-3. modify wp-config to point to new db info
-4. nav to theme dir
-5. npm install
-6. gulp
-7. touch a js file, touch a css file
-8. rip it
