--- _/admin/css/_editor.scss
@@ -1,5 +1,0 @@
-html {
-}
-
-body {
-}//body
