--- _/admin/css/style.scss
@@ -1,4 +1,0 @@
-@import '../../assets/css/normalize';
-@import '../../assets/css/variablesandmixins';
-@import '../../assets/css/customreset';
-@import 'editor';
