--- _/assets/css/style.scss
@@ -1,12 +1,0 @@
-/*!
-Theme Name: _blanket
-Theme URI:
-Author: Jonathan Brennan
-Description: Warm & Cozy.
-Text Domain: _blanket
-*/
-
-@import "variablesandmixins";
-@import "fonts";
-@import "normalize";
-@import "customreset";
