--- _/assets/js/entry.js
@@ -1,8 +1,0 @@
-//setup
-const body = document.body;
-
-//includes
-//=include utils.js
-//=include fetch.js
-//=include search.js
-//=include listeners.js
