+++ _/php/taxonomies.php
@@ -3,8 +3,7 @@
 // create taxonomies
 add_action('init', 'create_taxonomies', 0);
 
-function create_taxonomies()
-{
+function create_taxonomies() {
 
 	// News Category taxonomy, attached to news
   // $labels = array(
@@ -40,51 +39,5 @@ function create_taxonomies()
 
 } // end create taxonomies
 
-
-
-//create post types
-add_action('init', 'create_post_types');
-
-function create_post_types()
-{
-
-  // //create news post type
-  // $labels = array(
-  //   'name' => 'News',
-  //   'singular_name' => 'News Entry',
-  //   'add_new' => 'Add New',
-  //   'add_new_item' => 'Add New News Entry',
-  //   'edit_item' => 'Edit News Entry',
-  //   'new_item' => 'New News Entry',
-  //   'all_items' => 'All News Entries',
-  //   'view_item' => 'View News Entry',
-  //   'search_items' => 'Search News Entries',
-  //   'not_found' => 'No News Entries found',
-  //   'not_found_in_trash' => 'No News Entries found in Trash',
-  //   'parent_item_colon' => '',
-  //   'menu_name' => 'News Entries'
-  // );
-
-  // $args = array(
-  //   'labels' => $labels,
-  //   'exclude_from_search' => false,
-  //   'public' => true,
-  //   'publicly_queryable' => true,
-  //   'show_ui' => true,
-  //   'show_in_menu' => true,
-  //   'query_var' => true,
-  //   //'rewrite' => array('slug' => 'news'),
-  //   'capability_type' => 'post',
-  //   'has_archive' => true,
-  //   'hierarchical' => false,
-  //   'menu_position' => 5,
-  //   'menu_icon' => get_template_directory_uri() . '/_/assets/img/menuicon.png',
-  //   'supports' => array('title', 'thumbnail')
-  // );
-
-  // register_post_type('news', $args);
-
-
-} // end create post types
 
 ?>
