+++ _/php/theme.php
@@ -1,0 +1,16 @@
+<?php 
+//admin stuff
+// add_theme_support( 'post-thumbnails' ); 
+// set_post_thumbnail_size( 800, 450, true );
+// add_image_size('small', 384, 9999);
+// add_image_size('card', 800, 450, true);
+
+//options pages
+// if (function_exists('acf_add_options_page')) {
+//   acf_add_options_page(array(
+//     'page_title' => 'Navigation',
+//     'icon_url'   => get_template_directory_uri() . '/_/assets/img/menuicon.png',
+//     'position'   => 20
+//   ));
+// }
+?>
