+++ package.json
@@ -1,24 +1,42 @@
 {
   "name": "blanket",
-  "author": "jbrennan",
-  "version": "0.0.1",
-  "main": "/_/assets/js/app.js",
-  "devDependencies": {
-    "babel-core": "^6.26.3",
-    "babel-preset-env": "^1.7.0",
-    "babel-register": "^6.26.0",
-    "gulp": "^3.9.1",
-    "gulp-autoprefixer": "^3.1.0",
-    "gulp-babel": "^7.0.1",
-    "gulp-concat": "^2.5.2",
-    "gulp-include": "^2.3.0",
-    "gulp-line-ending-corrector": "^1.0.1",
-    "gulp-notify": "^2.2.0",
-    "gulp-rename": "^1.4.0",
-    "gulp-sass": "^2.2.0",
-    "gulp-sourcemaps": "^1.5.2",
-    "gulp-uglify": "^2.1.2",
-    "gulp-uglify-es": "^1.0.4"
+  "version": "2.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "build": "webpack --mode=production",
+    "dev": "webpack --watch"
   },
-  "dependencies": {}
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/jonathanawesome/blanket.git"
+  },
+  "authors": "Jonathan Brennan",
+  "devDependencies": {
+    "@babel/cli": "^7.1.5",
+    "@babel/core": "^7.1.6",
+    "@babel/preset-env": "^7.1.6",
+    "aurora-utilities": "^1.1.0",
+    "autoprefixer": "^9.3.1",
+    "babel-eslint": "^10.0.1",
+    "babel-loader": "^8.0.4",
+    "css-loader": "^1.0.1",
+    "css-mqpacker": "^7.0.0",
+    "mini-css-extract-plugin": "^0.4.4",
+    "node-sass": "^4.10.0",
+    "optimize-css-assets-webpack-plugin": "^5.0.1",
+    "postcss-loader": "^3.0.0",
+    "sanitize.scss": "^1.0.0",
+    "sass-loader": "^7.1.0",
+    "style-loader": "^0.23.1",
+    "uglifyjs-webpack-plugin": "^2.0.1",
+    "url-loader": "^1.1.2",
+    "webpack": "^4.25.1",
+    "webpack-cli": "^3.1.2"
+  },
+  "dependencies": {
+    "@babel/polyfill": "^7.0.0",
+    "promise-polyfill": "8.1.0",
+    "whatwg-fetch": "^3.0.0"
+  }
 }
