+++ src/editor.js
@@ -1,0 +1,4 @@
+/* blanket editor styles  */
+
+import './sass/editor';
+
