+++ src/sass/_fonts.scss
@@ -1,38 +1,38 @@
 
 @font-face {
   font-family: 'robotoregular';
-  src: url('_/assets/fonts/Roboto/Roboto-Regular-webfont.woff') format('woff');
+  src: url('../../_/fonts/Roboto/Roboto-Regular-webfont.woff') format('woff');
   font-weight: normal;
   font-style: normal;
 }
 @font-face {
   font-family: 'robotomedium';
-  src: url('_/assets/fonts/Roboto/Roboto-Medium-webfont.woff') format('woff');
+  src: url('../../_/fonts/Roboto/Roboto-Medium-webfont.woff') format('woff');
   font-weight: normal;
   font-style: normal;
 }
 @font-face {
   font-family: 'robotobold';
-  src: url('_/assets/fonts/Roboto/Roboto-Bold-webfont.woff') format('woff');
+  src: url('../../_/fonts/Roboto/Roboto-Bold-webfont.woff') format('woff');
   font-weight: normal;
   font-style: normal;
 }
 @font-face {
   font-family: 'robotoblack';
-  src: url('_/assets/fonts/Roboto/Roboto-Black-webfont.woff') format('woff');
+  src: url('../../_/fonts/Roboto/Roboto-Black-webfont.woff') format('woff');
   font-weight: normal;
   font-style: normal;
 }
 
 @font-face {
   font-family: 'roboto_condensedregular';
-  src: url('_/assets/fonts/Roboto/RobotoCondensed-Regular-webfont.woff') format('woff');
+  src: url('../../_/fonts/Roboto/RobotoCondensed-Regular-webfont.woff') format('woff');
   font-weight: normal;
   font-style: normal;
 }
 @font-face {
   font-family: 'roboto_condensedbold';
-  src: url('_/assets/fonts/Roboto/RobotoCondensed-Bold-webfont.woff') format('woff');
+  src: url('../../_/fonts/Roboto/RobotoCondensed-Bold-webfont.woff') format('woff');
   font-weight: normal;
   font-style: normal;
 }
