+++ src/sass/editor.scss
@@ -1,0 +1,9 @@
+@import "normalize";
+@import "variablesandmixins";
+@import "customreset";
+@import "fonts";
+
+
+body {
+  background-color: green;
+}
