+++ src/scripts/init.js
@@ -1,0 +1,13 @@
+(function () {
+
+  //variables
+  const body = document.body;
+
+
+  //When the DOM is ready, initialize all the things.
+  document.addEventListener('DOMContentLoaded', () =>
+    console.log("DOMContentLoaded")
+  );
+
+
+})();
