+++ webpack.config.js
@@ -1,0 +1,66 @@
+const path = require('path'),
+  MiniCssExtractPlugin = require('mini-css-extract-plugin'),
+  UglifyJSPlugin = require('uglifyjs-webpack-plugin'),
+  OptimizeCssAssetsPlugin = require('optimize-css-assets-webpack-plugin');
+
+module.exports = {
+  context: __dirname,
+  entry: {
+    main: './src/main.js',
+    editor: './src/editor.js',
+    login: './src/login.js'
+  },
+  output: {
+    path: path.resolve(__dirname, 'dist'),
+    filename: '[name]-bundle.js'
+  },
+  resolve: {
+    extensions: ['.js', '.scss', '.css', '.json']
+  },
+  mode: 'development',
+  devtool: 'cheap-eval-source-map',
+  module: {
+    rules: [
+      {
+        test: /\.js?$/,
+        loader: 'babel-loader'
+      },
+      {
+        test: /\.(png|woff|woff2|eot|ttf|svg)$/,
+        loader: 'url-loader?limit=100000'
+      },
+      {
+        test: /\.s?css$/,
+        use: [
+          MiniCssExtractPlugin.loader,
+          'css-loader',
+          {
+            loader: 'postcss-loader',
+            options: {
+              indent: 'postcss',
+              plugins: [
+                require('autoprefixer')({ browsers: 'last 2 versions' }),
+                require('css-mqpacker')({ sort: true })
+              ]
+            }
+          },
+          {
+            loader: 'sass-loader',
+            options: {
+              includePaths: [
+                'node_modules/sanitize.scss',
+                'node_modules/aurora-utilities/sass'
+              ]
+            }
+          }
+        ]
+      }
+    ]
+  },
+  plugins: [
+    new MiniCssExtractPlugin({ filename: '[name].css' }),
+  ],
+  optimization: {
+    minimizer: [new UglifyJSPlugin(), new OptimizeCssAssetsPlugin()]
+  }
+};
